moved all structs from staged-sync to config crate
