+++ crates/config/Cargo.toml
@@ -12,6 +12,7 @@ reth-staged-sync = { path = "../staged-sync" }
 reth-network = { path = "../net/network" }
 reth-net-nat = { path = "../../crates/net/nat" }
 reth-discv4 = { path = "../../crates/net/discv4" }
+reth-downloaders = { path = "../../crates/net/downloaders" }
 
 # io
 serde = "1.0"
